requirements.txt
yt-dlp>=2023.12.30
requests>=2.31.0
PyYAML>=6.0.1
colorama>=0.4.6
tqdm>=4.66.1